Refactor telegram bot command handler logic
