+++ GrammarNazi.Domain/Services/ITelegramCommandHandlerService.cs
@@ -1,0 +1,10 @@
+using System.Threading.Tasks;
+using Telegram.Bot.Types;
+
+namespace GrammarNazi.Domain.Services
+{
+    public interface ITelegramCommandHandlerService
+    {
+        Task HandleCommand(Message message);
+    }
+}
